Update Delete orphaned files

+++ src/dynfork/paladins/check.cr
@@ -166,27 +166,32 @@ module DynFork::QPaladins::Check
       file_val = nil
       curr_doc_hash = update? ? collection_ptr.value.find_one({_id: id}).not_nil!.to_h : nil
       {% for field in @type.instance_vars %}
-        unless @{{ field }}.ignored?
-          if !(file_val = curr_doc_hash[@{{ field }}.name]).nil?
-            if @{{ field }}.group == 4_u8 # FileField
-              if update?
-                # When updating the document.
-              else
-                # When creating a document.
-                if file_path = @{{ field }}.extract_file_path?
-                  File.delete(file_path.not_nil!)
-                  file_path = nil
-                end
+        if !@{{ field }}.ignored? && !@{{ field }}.value.nil?
+          if @{{ field }}.group == 4_u8 # FileField
+            if update?
+              # When updating the document.
+              if !(file_val = curr_doc_hash[@{{ field }}.name]).nil? 
+                file_path = file_val.not_nil!.as(DynFork::Globals::FileData).path
+                if file_path != curr_doc_hash
               end
-            elsif @{{ field }}.group == 5_u8 # ImageField
-              if update?
-                # When updating the document.
-              else
-                # When creating a document.
-                if img_dir_path = @{{ field }}.extract_images_dir_path?
-                  FileUtils.rm_rf(img_dir_path.not_nil!)
-                  img_dir_path = nil
-                end
+            else
+              # When creating a document.
+              if file_path = @{{ field }}.extract_file_path?
+                File.delete(file_path.not_nil!)
+                file_path = nil
+              end
+            end
+          elsif @{{ field }}.group == 5_u8 # ImageField
+            if update?
+              # When updating the document.
+              if !(file_val = curr_doc_hash[@{{ field }}.name]).nil?
+                # ???
+              end
+            else
+              # When creating a document.
+              if img_dir_path = @{{ field }}.extract_images_dir_path?
+                FileUtils.rm_rf(img_dir_path.not_nil!)
+                img_dir_path = nil
               end
             end
           end
